Add fun

+++ app/src/main/java/com/example/healt4u/model/Medicine.kt
@@ -26,6 +26,6 @@ data class Medicine(
     val expiredDate : Long ?=System.currentTimeMillis(),
     val afterEat :Boolean ?= true,
     val create_Date :Long ?= System.currentTimeMillis(),
-    val priority : Int ?= 0
+    val priority : Float ?= 0f
 )
